buildable in Linux with automake etc.
git-svn-id: svn+ssh://svn.luckybulldozer.com/srv/svn/luckybulldozer/mt2/MT2@198 30d454cc-f4c7-4f47-a1fd-a2c4f724ef71

+++ FistBlue/avatars/ryu.h
@@ -27,7 +27,7 @@ short sub_2d7d2(Player *ply);
 
 int RyuAtApex(Player *ply);
 
-short PLCBCompJumpRyuKen (Player *ply);	/* callback from 2c9be */
+int PLCBCompJumpRyuKen (Player *ply);	/* callback from 2c9be */
 void PLCBCompAttackRyuKen(Player *ply);
 short PLCBStandRyu(Player *ply);
 short PLCBCrouchRyu(Player *ply);
