+++ FistBlue/avatars/ryuken_comp.c
@@ -295,7 +295,7 @@ void PLCBCompAttackRyuKen(Player *ply) {
 	
 }
 
-short PLCBCompJumpRyuKen (Player *ply) {	/* callback from 2c9be */
+int PLCBCompJumpRyuKen (Player *ply) {	/* callback from 2c9be */
 	quirkysound(ply->ButtonStrength >> 1);
 	if(ply->VelX.full == 0) {
 		ply->Move = ply->ButtonStrength;
